REFACTOR: move SortContext.js
